DRY out views by moving more responsibilities up to CreateOrUpdateView
